add gas_collector getter & move unit test
no longer need to access private api
